# app/Http/Requests/Setting/UpdateSettingRequest.php
<?php

namespace App\Http\Requests\Setting;

use Illuminate\Foundation\Http\FormRequest;

/**
 * @OA\Schema(
 *     schema="UpdateSettingRequest",
 *     @OA\Property(property="company_name", type="string", maxLength=255, example="Nuevo Nombre S.A."),
 *     @OA\Property(property="company_description", type="string", maxLength=255, example="Nueva descripción"),
 *     @OA\Property(
 *         property="image",
 *         type="string",
 *         format="binary",
 *         description="Nueva imagen de logo (formatos: jpeg,png,jpg,gif,svg, tamaño máximo: 2MB)"
 *     )
 * )
 */
class UpdateSettingRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
     */
    public function rules(): array
    {
        return [
            'company_name' => 'sometimes|required|string|max:255',
            'company_description' => 'sometimes|required|string|max:255',
            'image' => 'sometimes|image|mimes:jpeg,png,jpg,gif,svg|max:2048'
        ];
    }
}