# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ricoton API",
        "description": "Documentación de la API de autenticación con Swagger en Laravel 12",
        "contact": {
            "email": "soporte@tuempresa.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/setting": {
            "get": {
                "tags": [
                    "Configuración"
                ],
                "summary": "Obtener configuración actual",
                "description": "Endpoint público para obtener la configuración de la empresa",
                "operationId": "3988a181344634d80af8d3dbf1d95179",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Setting"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Configuración"
                ],
                "summary": "Actualizar configuración existente",
                "description": "Endpoint protegido que requiere token de autenticación",
                "operationId": "b0691f1b7d0edf0e22edb8212f0f2165",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSettingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Configuración actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Settings updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Configuración no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Configuración"
                ],
                "summary": "Crear nueva configuración",
                "description": "Endpoint protegido que requiere token de autenticación",
                "operationId": "73e029d0273bbeb45ef7aac868663356",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSettingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Configuración creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Settings created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Setting"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Setting": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "company_name": {
                        "type": "string",
                        "example": "Mi Empresa"
                    },
                    "company_description": {
                        "type": "string",
                        "example": "Descripción de ejemplo"
                    },
                    "logo_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://ejemplo.com/storage/settings/logo.jpg"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "StoreSettingRequest": {
                "required": [
                    "company_name",
                    "company_description"
                ],
                "properties": {
                    "company_name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Mi Empresa S.A."
                    },
                    "company_description": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Descripción de la empresa"
                    },
                    "image": {
                        "description": "Imagen de logo (formatos: jpeg,png,jpg,gif,svg, tamaño máximo: 2MB)",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
            "UpdateSettingRequest": {
                "properties": {
                    "company_name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Nuevo Nombre S.A."
                    },
                    "company_description": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Nueva descripción"
                    },
                    "image": {
                        "description": "Nueva imagen de logo (formatos: jpeg,png,jpg,gif,svg, tamaño máximo: 2MB)",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}